Remove used QuizConfiguration + minor re-format
